Split JS cluster and super clusters tests and compile only on 1.16
